Memory block and using DStarLite with output of wavefront

+++ core/planning/structures/Grid.h
@@ -15,8 +15,8 @@ struct Grid {
   int rows;
   std::vector<GridCell> cells;
   Grid() : columns(GRID_WIDTH), rows(GRID_HEIGHT), cell_width(CELL_WIDTH), cell_height(CELL_HEIGHT),width(FIELD_WIDTH),height(FIELD_HEIGHT) {
-    for (int r = 0; r < std::round(height/cell_height); ++r) {
-      for (int c = 0; c < std::round(width/cell_width); ++c) {
+    for (int r = 0; r < std::round((float)height/(float)cell_height); ++r) {
+      for (int c = 0; c < std::round((float)width/(float)cell_width); ++c) {
         cells.push_back(GridCell(r,c));
       }
     }
